added screenshots added allure.attach added separate pages in d3a.io added reporting using allure-behave

+++ features/steps/d3asimulationsteps.py
@@ -1,12 +1,13 @@
 from behave import *
-from selenium import webdriver
-#from d3aBehaveProjectTestSuite.venv.commons import loginpage
-from common import *
+from features.pages.projectspage import *
 import time
+import pyautogui
+import allure
+from allure_commons.types import AttachmentType
 
-@given('the user creates a project')
-def create_project_call(context):
-    create_project(context, "simproject", "simdescription")
+@given('the user creates a project with Project Name as "{name}" and description as "{description}"')
+def create_project_call(context, name, description):
+    create_project(context, name, description)
 
 @given('the user selects on the project')
 def select_project_call(context):
@@ -23,38 +24,47 @@ def enter_simulation_details_call(context, sname, sdescription):
 @when('the user clicks on Next button')
 def click_next_button_call(context):
     click_next_button(context)
-
-@then('a new simulation is created')
-def new_simulation_validation(context):
-    modelling = context.driver.find_element_by_xpath('//*[@id="root"]/div/div[2]/header/h1').text
-    assert modelling == "Modelling"
-    print("***New simulation has been created - Landed on modelling page***")
-
-@then('is visible on projects page under the project')
-def new_simulation_visible_validation(context):
     time.sleep(2)
+
+@then('a new simulation is visible on projects page under the project')
+def new_simulation_visible_validation(context):
     load_project_page_url(context)
     select_project(context)
-    newsimulation = context.driver.find_element_by_xpath('//*[@id="root"]/div/div[2]/div[2]/div/div/section/div[1]/div[2]/div/div[1]/div/div[1]/div[1]/div/p').text
-    assert newsimulation == "sampleSimulationName"
-    print(newsimulation + "***New simulation has been created - Visible on projects page***")
-    time.sleep(1)
-    delete_project(context)
-    context.driver.close()
+    try:
+        newSimulationName = new_simulation_validation(context)
+    except:
+        context.driver.close()
+        assert False, "Test Failed"
+    if new_simulation_validation(context) == "sampleSimulationName":
+        print("***New simulation has been created - Visible on projects page***")
+        time.sleep(1)
+        assert True, "Test Passed"
+        myScreenshot = pyautogui.screenshot()
+        myScreenshot.save(r'C:\Users\lakshmi krishnaswamy\screenshots\new_simulation_visible_validation_new_project.png')
+        allure.attach(context.driver.get_screenshot_as_png(), name="new_simulation_visible_validation_new_project",attachment_type=AttachmentType.PNG)
+        #delete_project(context)
+        context.driver.close()
 
 @given('the user selects an existing project')
 def select_project_call(context):
     load_project_page_url(context)
     select_project(context)
 
-@then('is visible on projects page under the existing project')
+@then('a new simulation is visible on projects page under the existing project')
 def new_simulation_visible_validation(context):
-    time.sleep(2)
     load_project_page_url(context)
     select_project(context)
-    newsimulation = context.driver.find_element_by_xpath('//*[@id="root"]/div/div[2]/div[2]/div/div/section/div[1]/div[2]/div/div[1]/div/div[1]/div[1]/div/p').text
-    assert newsimulation == "sampleSimulationName"
-    print(newsimulation + "***New simulation has been created - Visible on projects page***")
-    time.sleep(1)
-    delete_simulation(context)
-    context.driver.close()
+    try:
+        newSimulationName = new_simulation_validation(context)
+    except:
+        context.driver.close()
+        assert False, "Test Failed"
+    if new_simulation_validation(context) == "SimulationNameExistingProj":
+        print("***New simulation has been created - Visible on projects page***")
+        time.sleep(1)
+        assert True, "Test Passed"
+        myScreenshot = pyautogui.screenshot()
+        myScreenshot.save(r'C:\Users\lakshmi krishnaswamy\screenshots\new_simulation_visible_validation_existing_project.png')
+        allure.attach(context.driver.get_screenshot_as_png(), name="new_simulation_visible_validation_existing_project", attachment_type=AttachmentType.PNG)
+        delete_project(context)
+        context.driver.close()
